+++ src/main/java/com/nnk/springboot/repositories/BidListRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BidListRepository extends JpaRepository<BidList, Long> {
+public interface BidListRepository extends JpaRepository<BidList, Integer> {
 
 }
