+++ src/test/java/com/nnk/springboot/repository/BidTests.java
@@ -37,7 +37,7 @@ public class BidTests {
 		Assertions.assertTrue(listResult.size() > 0);
 
 		// Delete
-		Long id = bid.getId();
+		Integer id = bid.getId();
 		bidListRepository.delete(bid);
 		Optional<BidList> bidList = bidListRepository.findById(id);
 		Assertions.assertFalse(bidList.isPresent());
