+++ src/test/java/com/nnk/springboot/repository/RatingTests.java
@@ -36,7 +36,7 @@ public class RatingTests {
 		Assertions.assertTrue(listResult.size() > 0);
 
 		// Delete
-		Long id = rating.getId();
+		Integer id = rating.getId();
 		ratingRepository.delete(rating);
 		Optional<Rating> ratingList = ratingRepository.findById(id);
 		Assertions.assertFalse(ratingList.isPresent());
