+++ src/test/java/com/nnk/springboot/repository/RuleTests.java
@@ -37,7 +37,7 @@ public class RuleTests {
 		Assertions.assertTrue(listResult.size() > 0);
 
 		// Delete
-		Long id = rule.getId();
+		Integer id = rule.getId();
 		ruleNameRepository.delete(rule);
 		Optional<RuleName> ruleList = ruleNameRepository.findById(id);
 		Assertions.assertFalse(ruleList.isPresent());
